add storybook and break up UI into smaller components
