+++ frontend/desktop/.gitignore
@@ -1,5 +1,6 @@
 node_modules/
 dist/
 release/
+storybook-static/
 src-tauri/target/
 *.log
