+++ frontend/desktop/package.json
@@ -10,6 +10,9 @@
     "dev:desktop": "node scripts/tauri.mjs dev",
     "dev": "vite --host 127.0.0.1",
     "build": "tsc --noEmit && vite build",
+    "storybook": "storybook dev -p 6006",
+    "build-storybook": "storybook build",
+    "test-storybook": "vitest run --project=storybook",
     "start": "node scripts/tauri.mjs dev",
     "tauri": "node scripts/tauri.mjs",
     "dist": "node scripts/tauri.mjs build",
@@ -27,12 +30,19 @@
     "react-dom": "^19.2.8"
   },
   "devDependencies": {
+    "@storybook/addon-a11y": "^10.5.9",
+    "@storybook/addon-vitest": "^10.5.9",
+    "@storybook/react-vite": "^10.5.9",
     "@tauri-apps/cli": "^2.11.4",
     "@types/node": "^26.1.2",
     "@types/react": "^19.2.18",
     "@types/react-dom": "^19.2.4",
     "@vitejs/plugin-react": "^6.0.5",
+    "@vitest/browser-playwright": "^4.1.11",
+    "playwright": "^1.62.1",
+    "storybook": "^10.5.9",
     "typescript": "^7.0.2",
-    "vite": "^8.2.0"
+    "vite": "^8.2.0",
+    "vitest": "^4.1.11"
   }
 }
