+++ frontend/desktop/src/styles.css
@@ -8,7 +8,7 @@
   text-rendering: optimizeLegibility;
   --ink: #1f2328;
   --muted: #656d76;
-  --subtle: #8c959f;
+  --subtle: #6b737c;
   --border: #d8dee4;
   --border-soft: #eaeef2;
   --surface: #f6f8fa;
@@ -64,6 +64,10 @@ p {
   min-height: 100vh;
   display: grid;
   grid-template-columns: 196px minmax(0, 1fr);
+}
+
+.desktop-shell.enrollment-shell {
+  display: block;
 }
 
 .desktop-sidebar {
@@ -249,6 +253,32 @@ p {
   border-color: #e8c8c2;
 }
 
+.notice + .notice {
+  margin-top: -16px;
+}
+
+.status-loading {
+  min-height: 260px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  gap: 12px;
+  color: var(--accent);
+}
+
+.status-loading h2 {
+  margin: 0;
+  color: var(--ink);
+  font-size: 16px;
+  font-weight: 600;
+}
+
+.status-loading p {
+  margin: 4px 0 0;
+  color: var(--muted);
+  font-size: 12px;
+}
+
 .page-stack {
   display: flex;
   flex-direction: column;
@@ -262,7 +292,7 @@ p {
   justify-content: space-between;
 }
 
-.page-heading h1 {
+.page-heading h2 {
   margin: 0;
   color: var(--ink);
   font-size: 20px;
@@ -318,7 +348,7 @@ p {
   margin: 0 0 10px;
 }
 
-.enrollment-welcome h1 {
+.enrollment-welcome h2 {
   margin: 0;
   color: var(--ink);
   font-size: 28px;
@@ -337,6 +367,17 @@ p {
 .enrollment-action {
   min-width: 150px;
   min-height: 38px;
+}
+
+.enrollment-progress {
+  min-height: 38px;
+  padding: 0 12px;
+  display: inline-flex;
+  align-items: center;
+  gap: 8px;
+  color: var(--accent);
+  font-size: 12px;
+  font-weight: 650;
 }
 
 .enrollment-welcome small {
@@ -375,7 +416,7 @@ p {
   margin: 0 0 5px;
 }
 
-.status-hero h1 {
+.status-hero h2 {
   margin: 0;
   color: var(--ink);
   font-size: 22px;
@@ -618,8 +659,8 @@ p {
 
 .switch {
   position: relative;
-  width: 38px;
-  height: 22px;
+  width: 44px;
+  height: 28px;
   flex: 0 0 auto;
   cursor: pointer;
 }
@@ -644,8 +685,8 @@ p {
   position: absolute;
   top: 3px;
   left: 3px;
-  width: 16px;
-  height: 16px;
+  width: 22px;
+  height: 22px;
   background: #fff;
   border-radius: 50%;
   box-shadow: 0 1px 2px rgb(22 28 25 / 20%);
@@ -765,6 +806,22 @@ p {
   font-size: 12px;
 }
 
+.empty-inline > div {
+  max-width: 470px;
+  display: flex;
+  flex-direction: column;
+  gap: 4px;
+}
+
+.empty-inline strong {
+  color: var(--ink);
+  font-size: 12px;
+}
+
+.empty-inline span {
+  line-height: 1.45;
+}
+
 .page-error {
   min-height: 260px;
   padding: 48px 32px;
@@ -830,6 +887,83 @@ p {
 
   .stat-card:last-child {
     border-bottom: 0;
+  }
+}
+
+@media (max-width: 520px) {
+  .desktop-shell {
+    display: block;
+  }
+
+  .desktop-sidebar {
+    position: static;
+    width: 100%;
+    min-height: 56px;
+    padding: 8px 12px;
+    flex-direction: row;
+    align-items: center;
+    justify-content: space-between;
+    border-right: 0;
+    border-bottom: 1px solid var(--border);
+  }
+
+  .desktop-brand {
+    height: 40px;
+    padding: 0 4px;
+  }
+
+  .desktop-brand span {
+    display: none;
+  }
+
+  .desktop-nav {
+    margin-top: 0;
+    flex-direction: row;
+    gap: 4px;
+  }
+
+  .desktop-nav button {
+    min-height: 40px;
+    padding: 0 10px;
+  }
+
+  .desktop-main {
+    grid-column: auto;
+  }
+
+  .desktop-page-header {
+    height: auto;
+    min-height: 56px;
+    padding: 10px 16px;
+    gap: 12px;
+  }
+
+  .desktop-content {
+    padding: 24px 16px 48px;
+  }
+
+  .page-heading {
+    padding-right: 0;
+  }
+
+  .status-row {
+    grid-template-columns: 34px minmax(0, 1fr);
+    align-items: start;
+  }
+
+  .status-row > .badge {
+    grid-column: 2;
+    justify-self: start;
+  }
+
+  .advanced-config-heading,
+  .advanced-danger-zone {
+    align-items: flex-start;
+    flex-direction: column;
+  }
+
+  .logout-confirmation {
+    align-items: flex-start;
   }
 }
