+++ frontend/desktop/tsconfig.json
@@ -3,5 +3,5 @@
   "compilerOptions": {
     "types": ["vite/client", "node"]
   },
-  "include": ["src", "vite.config.ts"]
+  "include": ["src", ".storybook", "vite.config.ts", "vitest.config.ts"]
 }
