+++ frontend/package.json
@@ -4,9 +4,14 @@
   "packageManager": "pnpm@11.3.0",
   "scripts": {
     "build": "pnpm --filter @agentdesktop/controller-web build && pnpm --filter @agentdesktop/desktop-web build",
-    "check": "pnpm --filter @agentdesktop/controller-web check && pnpm --filter @agentdesktop/controller-web exec biome check ../desktop && pnpm --filter @agentdesktop/desktop-web build",
+    "check": "pnpm --filter @agentdesktop/controller-web check && pnpm --filter @agentdesktop/controller-web exec biome check ../desktop && pnpm --filter @agentdesktop/desktop-web build && pnpm test:storybook",
     "dev:controller": "pnpm --filter @agentdesktop/controller-web dev",
     "dev:desktop": "pnpm --filter @agentdesktop/desktop-web dev:desktop",
+    "storybook": "pnpm --parallel --filter @agentdesktop/desktop-web --filter @agentdesktop/controller-web run storybook",
+    "storybook:desktop": "pnpm --filter @agentdesktop/desktop-web storybook",
+    "storybook:controller": "pnpm --filter @agentdesktop/controller-web storybook",
+    "build:storybook": "pnpm --filter @agentdesktop/desktop-web build-storybook && pnpm --filter @agentdesktop/controller-web build-storybook",
+    "test:storybook": "pnpm --filter @agentdesktop/desktop-web test-storybook && pnpm --filter @agentdesktop/controller-web test-storybook",
     "format": "pnpm --filter @agentdesktop/controller-web format && pnpm --filter @agentdesktop/controller-web exec biome format --write ../desktop"
   }
 }
