+++ frontend/pnpm-workspace.yaml
@@ -2,3 +2,16 @@ packages:
   - controller
   - desktop
   - ui
+allowBuilds:
+  esbuild: true
+minimumReleaseAgeExclude:
+  - '@vitest/browser-playwright@4.1.11'
+  - '@vitest/browser@4.1.11'
+  - '@vitest/expect@4.1.11'
+  - '@vitest/mocker@4.1.11'
+  - '@vitest/pretty-format@4.1.11'
+  - '@vitest/runner@4.1.11'
+  - '@vitest/snapshot@4.1.11'
+  - '@vitest/spy@4.1.11'
+  - '@vitest/utils@4.1.11'
+  - vitest@4.1.11
