+++ frontend/ui/src/styles.css
@@ -6,7 +6,7 @@
   background: #fff;
   font-synthesis: none;
   --muted: #656d76;
-  --subtle: #8c959f;
+  --subtle: #6b737c;
   --border: #d8dee4;
   --border-soft: #eaeef2;
   --surface: #f6f8fa;
@@ -77,17 +77,38 @@
   min-height: 57px;
   padding: 10px 15px;
   display: grid;
-  grid-template-columns: minmax(140px, 0.8fr) 90px minmax(220px, 1.5fr) auto;
+  grid-template-columns:
+    minmax(140px, 0.8fr) 90px minmax(220px, 1.5fr) auto
+    18px;
   align-items: center;
   gap: 14px;
   color: #4b535c;
   font-size: 12px;
   cursor: pointer;
-  list-style-position: outside;
+  list-style: none;
+}
+
+.tool-inventory-item summary::-webkit-details-marker {
+  display: none;
 }
 
 .tool-inventory-item summary:hover {
   background: #fafbfc;
+}
+
+.tool-inventory-item summary:focus-visible,
+.mini-pager button:focus-visible {
+  outline: 3px solid rgb(95 67 178 / 18%);
+  outline-offset: -3px;
+}
+
+.tool-disclosure {
+  color: var(--muted);
+  transition: transform 150ms ease;
+}
+
+.tool-inventory-item[open] > summary .tool-disclosure {
+  transform: rotate(90deg);
 }
 
 .tool-cell strong {
@@ -148,8 +169,8 @@
 }
 
 .mini-pager button {
-  width: 19px;
-  height: 19px;
+  width: 28px;
+  height: 28px;
   padding: 0;
   display: grid;
   place-items: center;
@@ -218,10 +239,29 @@
 
 @media (max-width: 800px) {
   .tool-inventory-item summary {
-    grid-template-columns: minmax(130px, 1fr) auto;
+    grid-template-columns: minmax(130px, 1fr) auto 20px;
+  }
+
+  .tool-cell {
+    grid-column: 1;
+    grid-row: 1;
+  }
+
+  .tool-version {
+    grid-column: 2;
+    grid-row: 1;
+  }
+
+  .tool-disclosure {
+    grid-column: 3;
+    grid-row: 1;
   }
 
   .tool-path {
+    grid-column: 1 / -1;
+  }
+
+  .capability-counts {
     grid-column: 1 / -1;
   }
